requirements.txt
discord.py==1.5.1
iservscrapping==4.1.3
pymongo~=3.11.0